services/alert_ingestion/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
pydantic==2.5.3
prometheus-client==0.19.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis[hiredis]==5.0.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
